Playing With ADC and Digits
I'm beginning to experiment with the ADC and a couple of WS2812B
digits.  I only have two for now, so I'm just testing using
potentiometers to set the minute components of the work and rest
times.

It wasn't feasible to cram a decoder for a more memory-efficient type
of storage into the digit writing routine, so I just went ahead and
switched to an atmega324p as the controller for now because I had one
on hand.  For the actual production clock I'll use a 328p, but no one
has them in stock in DIP form at the moment.