Counting Down
This gets me to a point where the timer actually, well, times, albeit
with only two digits.  I've got the 16 bit counter in action to count
off the seconds and rearranged the main loop to work as a state
machine with some global state to make it all work with the timer
interrupt.