Firmware Cleanup
I've got all the test hardware set up now, so at this point it's time
to stop toying around with demo code and start writing a serious
firmware.  To that end, I made a handful of enhancements here that
shouldn't materially affect the behavior:

* Added missing copyright headers
* Defined constants for all of the I/O pins
* Added a `setup` function instead of dumping all the setup in `main`
* Combined `debounce_start` and `debounce_stop` into a single
  `debounce_button` function that handles the button I/O as well
* Pulled ADC reading out into a separate function
* Removed duplicate `WORK`/`REST` state logic with a fallthrough in
  the switch

The main loop also seemed to be running a little too fast to give the
WS2812B LEDs on my new digit boards enough reset time between
refreshes, so I added a 1ms delay to get that displaying properly.