Temporary Rest Time Display
Previously, when the timer wasn't active we would just show whatever
time was most recently altered.  This changes it so that we always
show the work time by default, since that's the time that will take
effect right away when the user pushes the start button.  The rest
time will show while the user is adjusting it, and then for five
seconds afterwards.

This uses timer 0 to count off the time for the display separately
from the primary timer.  I also added a `set_number` function to
display a four-digit decimal number across the display, which I used
for debugging this feature and later will make its way into a
diagnostic mode for profiling potentiometers.