Buzzer Implementation
This replaces the old, somewhat janky buzzer implementation with
something closer to how I want the final product to behave with an
alternating buzz for starting the work time and a solid one for
starting the rest time.  I also handled the edge cases when the timer
runs with either the rest or the work time set to zero, in which case
we just repeat the one time that is set over and over again.