Adjustable Brightness
This adds an ADC input (hooked up to a potentiometer) to set the
display brightness in five separate levels.  For the debounce function
I manually set a series of thresholds, which is a model I'll be moving
all of the ADC debounce functions to in the next couple of commits as
it enables profiling a potentiometer with a non-linear response.  I
also noticed that I'd accidentally left some `else if (0)` blocks in
the other debounce functions from debugging a while back and removed
them, which has improved the stability of the minute and second
potentiometers as well.