Added Diagnostic Mode
This adds a diagnostic mode to the firmware.  By connecting a pin to
ground (which will be a header + jumper on the final control board)
you can put it into a mode where it just draws the raw value of the
work minute pot to the screen. This will make it possible to profile
potentiometers to compensate for any non-linearity