Debounce Refactor and RF Button Addition
This adds pins for and checks an RF receiver for use as a remote
control.  Along the way I refactored the button debouncing so that
buttons will now fire once and not fire again until they've fully
reset.  This is used for all buttons now and obviates the check for
button release after pause.