Added Remote PCB
This adds a simple little PCB design to mount a QIACHIP 433Mhz
transmitter and connect it to a couple of buttons and an antenna.  I
used through-hole resistors in this board because it just didn't make
sense to pay for SMT assembly when the two resistors are the only
components on the board.