Fix Discontinuities in Enclosure Cut File
The FreeCAD export I used to lay out the cut paths for the seven
segment digits had a bunch of open paths for the actual segments.  For
some reason the Origin *mostly* took these paths in stride, but a few
of them would show up as on line cuts instead of inside cuts on the
device and need manual correction.  This fixes the paths so they
should cut correctly by default.