Rotated Remote Top
This includes a new STL file for the top component of the remote that
doesn't require manual rotation when slicing.  At this point the
components all appear to fit in the remote, so I think this is going
to be the final design.  Ideally I'd like to tweak the battery
compartment a little more, but I've made such a mess referencing
external geometry that it's just too difficult to make those kinds of
changes at this point.