Remove Remote Screws
Since it's so easy to break off the stems by overtightening the remote
screws, I'm removing the holes for them from the back.  The board will
still screw down to the top piece, but sealing up the shell will be
done with super glue now.  This isn't ideal for repairability, but
should hold up better in practice.